requirements.txt
osmnx>=1.9.1
numpy>=1.26.0
shapely>=2.0.2
trimesh[easy]>=4.0.5
geopandas>=0.14.1
networkx>=3.2.1
pyproj>=3.6.1
click>=8.1.7
scipy>=1.11.4
aiohttp>=3.9.1
python-dotenv>=1.0.0
scikit-learn>=1.3.2
tqdm>=4.66.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
geopy>=2.4.1
pygltflib>=1.16.1
overturemaps>=0.9.0
Pillow>=10.0.0
rasterio>=1.3.0
